Fix test UB: signed overflow in interleaved pseudo-random (GCC exposed)
(round * 104729) overflows int32 at round >= ~20.5k. MSVC wraps silently;
GCC -O2 exploits the UB producing garbage max_count, which corrupted the
stack and made the interleaved test fail with bogus timestamps. Switch to
unsigned arithmetic.

Also: Linux smoke-build now verified in Docker (gcc:13 + libevdev/libudev):
- cmake configure/build clean
- spsc_test 6/6 Release runs PASS
- ASan+UBSan run PASS

+++ tests/spsc_test.cpp
@@ -126,10 +126,10 @@ static int test_interleaved_overflow() {
     bool first = true;
     HidSample out[64];
     for (int round = 0; round < 50000; ++round) {
-        const int pushes = 1 + (round * 7919) % 20;   // deterministic pseudo-random
+        const int pushes = 1 + static_cast<int>((round * 7919u) % 20u);  // unsigned: no UB
         for (int p = 0; p < pushes; ++p)
             ring.push(make(produced++));
-        const size_t n = ring.pop_batch(out, 1 + (round * 104729) % 32);
+        const size_t n = ring.pop_batch(out, 1 + (unsigned)(round * 104729u) % 32u);
         for (size_t i = 0; i < n; ++i) {
             if (!first && out[i].timestamp <= last) {
                 fprintf(stderr, "FAIL interleaved: %u after %u\n",
